Hangman-Game
I've build a responsive Hangman game using HTML, CSS, and JavaScript. In this game, the user has to guess all the letters of a randomly generated word within a given number of tries. There is also a hangman illustration that will progressively appear on the gallows for each incorrect guess.